src/main/webapp/app/entities/establecimientos/zona/delete/zona-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ZonaService } from '../service/zona.service';

import { ZonaDeleteDialogComponent } from './zona-delete-dialog.component';

describe('Zona Management Delete Component', () => {
  let comp: ZonaDeleteDialogComponent;
  let fixture: ComponentFixture<ZonaDeleteDialogComponent>;
  let service: ZonaService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, ZonaDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(ZonaDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(ZonaDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(ZonaService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
